compare_build.sh: Add msg for git checkout failure
This is important when you have uncommited changes that
could be clobbered.

EXAMPLE OUTPUT
Failed to checkout c1592fff1:
error: Your local changes to the following files would be overwritten by checkout:
	Makefile.rules
Please commit your changes or stash them before you switch branches.
Aborting

BRANCH=none
BUG=none
TEST=# Modify a line of code near a change in HEAD
     ./util/compare_build.sh

+++ util/compare_build.sh
@@ -51,7 +51,14 @@ do_build() {
   local ref="$1"
   local result_dir="$2"
   local board="$3"
-  git checkout "${ref}" >/dev/null 2>&1
+
+  local err
+  if ! err=$(git checkout "${ref}" 2>&1); then
+    echo "Failed to checkout ${ref}:" >&2
+    echo "${err}" >&2
+    echo >&2
+    return 1
+  fi
   echo "Testing commit: $(git rev-parse --short HEAD)"
   rm -rf "build/${board}"
   # STATIC_VERSION makes sure the generated ec_version.h is constant.
